src/sglang_deployment.py
"""SGLang deployment example for multimodal pipelines."""


def demonstrate_sglang_deployment():
    """Demonstrate SGLang deployment concepts."""

    print("SGLang: Advanced Model Serving for Multimodal Pipelines")
    print("=" * 50)

    print("\nNote: This is a conceptual demonstration.")
    print("For full SGLang deployment, install: pip install 'sglang[all]'")

    # Conceptual pipeline example
    print("\n1. Example Multimodal Pipeline:")
    print(
        """
Customer Support Automation Pipeline:

Input:
    ├── Screenshot (Image)
    └── Voice Message (Audio)
            ↓
Processing Graph:
    ├── Image Classifier Node
    │   └── Classify: error/feature_request/other
    ├── Audio Transcriber Node
    │   └── Convert speech to text
    └── → Combiner Node
            ↓
    Summary Generator Node
            ↓
Output:
    └── Support Ticket Summary
"""
    )

    print("\n2. SGLang Benefits:")
    print("   ✓ Graph-based pipeline definition")
    print("   ✓ Automatic batching and optimization")
    print("   ✓ Support for quantization (FP8/INT4/AWQ)")
    print("   ✓ Multi-LoRA serving")
    print("   ✓ Streaming and async processing")

    print("\n3. Sample SGLang Pipeline Code:")
    print(
        """
import sglang as sgl

@sgl.function
def classify_image(s, image):
    s += sgl.image(image)
    s += "Classify this support screenshot: "
    s += sgl.gen("classification", max_tokens=10)

@sgl.function
def transcribe_audio(s, audio):
    # Use speech model
    s += "Transcribed: " + speech_to_text(audio)

@sgl.function
def summarize_ticket(s, img_class, transcript):
    s += f"Image type: {img_class}\\n"
    s += f"User said: {transcript}\\n"
    s += "Summary: " + sgl.gen("summary", max_tokens=100)
"""
    )

    print("\n4. Deployment Options:")
    print("   - Local server: python -m sglang.launch_server")
    print("   - Docker container for isolation")
    print("   - Kubernetes for scale")
    print("   - Cloud endpoints (AWS, GCP, Azure)")

    print("\n5. Performance Optimizations:")
    print("   - Quantization reduces memory 4x")
    print("   - RadixAttention for KV cache sharing")
    print("   - Speculative decoding for 2-3x speedup")
    print("   - Continuous batching for throughput")

    print("\nFor production deployment:")
    print("1. Define your pipeline as SGLang functions")
    print("2. Configure quantization and optimization")
    print("3. Launch server with your pipeline")
    print("4. Integrate with your application via REST API")

    print("\nLearn more: https://github.com/sgl-project/sglang")


if __name__ == "__main__":
    print("=== SGLang Deployment Example ===\n")
    demonstrate_sglang_deployment()
