pyproject.toml
[tool.poetry]
name = "multimodal-ai"
version = "0.1.0"
description = "Multimodal AI with Vision, Audio, and Generative Models - Working Examples"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
transformers = "^4.40.0"
torch = ">=2.6.0"
torchvision = ">=0.21.0"
torchaudio = ">=2.6.0"
diffusers = "^0.30.0"
accelerate = "^0.25.0"
pillow = "^10.2.0"
soundfile = "^0.12.1"
librosa = "^0.10.1"
opencv-python = "^4.9.0"
matplotlib = "^3.8.0"
seaborn = "^0.13.0"
faiss-cpu = "^1.7.4"
sentencepiece = "^0.2.0"
python-dotenv = "^1.0.0"
pandas = "^2.1.0"
numpy = "^1.26.0"
requests = "^2.31.0"
gradio = "^4.16.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
black = "^24.0.0"
ruff = "^0.6.0"
jupyter = "^1.0.0"
ipykernel = "^6.29.0"
ipywidgets = "^8.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py312']

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "C4", "SIM"]
